[FEATURE] Improve junction detection from SA tag iGenVar.hpp/cpp - add CLI parameter (max_overlap) for max. ovl. between aln. segments - make min_var_length and max_overlap uint32_t
analyze_sa_tag_method.cpp:
- use max_overlap instead of hard-coded value
- enforce minimum variant length
- produce junctions when distance on read is large (insertions)
- handle slightly overlapping alignment segments
- pass args to analyze_sa_tag()

+++ include/modules/sv_detection_methods/analyze_cigar_method.hpp
@@ -36,4 +36,4 @@ void analyze_cigar(std::string const & read_name,
                    std::vector<seqan3::cigar> & cigar_string,
                    seqan3::dna5_vector const & query_sequence,
                    std::vector<Junction> & junctions,
-                   uint64_t const min_length);
+                   uint32_t const min_length);
